feat: add format function support
This commit removes the `formatForSelenium` config option with a
`format` option. That key accepts a function or `false`. The fuction
will recieve one argument, `capabilites` array, wich it can then use to
create any output. The purpose of this change is to allow users to
export the results in any format they wish.

The `format` functionality defailts to `formatForSelenium`'s previous
behavior. It does this by using a helper function of the same name to
format the `capabilities` array. The helper function is exported by the
package.